# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The response body of every refusal this API makes.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Refusal"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "description": "What kind of refusal this is, as something a client can branch on.\n\nCodes rather than status alone, because 422 is doing two jobs \u2014 a malformed request\nbody and a panel that breaks the input contract are both 422 and are not the same\nproblem. A client retries the first after fixing its JSON and the second after\nfixing its data.",
        "enum": [
          "tenant_required",
          "invalid_request",
          "contract_violation",
          "payload_rejected",
          "not_found",
          "in_use",
          "artefact_unavailable",
          "too_large",
          "engine_refused",
          "internal"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "Estimator": {
        "description": "How Mode A's numbers are arrived at. **Not** which mode or rung is used.\n\nThe distinction matters, because ``assess`` deliberately exposes no way to force a\nmode or a rung and a test asserts it never grows one. That rule is about *data\nadequacy*: whether a panel can support seven terms or three is the engine's call,\nnever the caller's, because a caller who could overrule it would overrule it.\n\nThis is a different question. The ladder still decides the mode, the rung and the\nterms, identically either way; this only decides how the coefficients on those\nterms are estimated. A test pins that: the same panel returns the same mode, rung\nand factor list under both estimators.",
        "enum": [
          "nb2",
          "bayes"
        ],
        "title": "Estimator",
        "type": "string"
      },
      "FacilityType": {
        "description": "The road type a weight was estimated on, or that a corridor is.\n\n``ANY`` on a weight means the source does not restrict by facility. ``ANY`` on a\nrun means the corridor type was not declared, in which case only unrestricted\nweights are admissible \u2014 the engine will not guess.",
        "enum": [
          "rural_two_lane",
          "rural_multilane",
          "urban_arterial",
          "any"
        ],
        "title": "FacilityType",
        "type": "string"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "description": "One thing wrong with the request, and where.",
        "properties": {
          "location": {
            "description": "Dotted path into the request, e.g. 'body.name'.",
            "title": "Location",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Job": {
        "additionalProperties": false,
        "description": "One request to assess a corridor, and where it got to.\n\n`params` is the assessment's own options \u2014 estimator, priors, spatial, which\nadapters to run. Stored as given so that a job can be re-run identically, and so\nthat the manifest's fingerprint has something to be checked against.",
        "properties": {
          "corridor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corridor Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "project_id"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobOptions": {
        "additionalProperties": false,
        "description": "The assessment's own options \u2014 exactly the CLI's, and nothing invented here.\n\nEverything absent has the same default the command line has, so a job submitted\nwith no options at all is the run `roadrisk corridor` would have done.\n\nWhat is deliberately **not** here: any way to force a mode, a rung or a term. The\nengine decides those from data adequacy, `assess` exposes no argument for them, and\na test in `tests/test_engine.py` asserts it never grows one. A caller who could\noverrule the ladder would overrule it.",
        "properties": {
          "adapters": {
            "items": {
              "enum": [
                "osm",
                "rasters",
                "traffic",
                "mapillary"
              ],
              "type": "string"
            },
            "title": "Adapters",
            "type": "array"
          },
          "estimator": {
            "$ref": "#/components/schemas/Estimator",
            "default": "nb2"
          },
          "facility_type": {
            "$ref": "#/components/schemas/FacilityType",
            "default": "any"
          },
          "n_periods": {
            "default": 24,
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "N Periods",
            "type": "integer"
          },
          "region": {
            "$ref": "#/components/schemas/Region",
            "default": "global"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "default": "all"
          },
          "shape_factors": {
            "description": "Factors to put a rung 3 spline on. Validated against `factors.yaml` at submit \u2014 a name no factor has is a typo, and finding it in a run log a quarter of an hour later helps nobody.",
            "items": {
              "type": "string"
            },
            "title": "Shape Factors",
            "type": "array"
          },
          "tolerance_m": {
            "default": 30.0,
            "exclusiveMinimum": 0.0,
            "maximum": 1000.0,
            "title": "Tolerance M",
            "type": "number"
          },
          "unit_length_m": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Length M"
          },
          "use_registry_priors": {
            "default": false,
            "title": "Use Registry Priors",
            "type": "boolean"
          },
          "use_spatial": {
            "default": false,
            "title": "Use Spatial",
            "type": "boolean"
          }
        },
        "title": "JobOptions",
        "type": "object"
      },
      "JobStatus": {
        "description": "Where a job is, and nothing about what it concluded.\n\nThe distinction this vocabulary exists to protect: **a job that descended to Mode B\nsucceeded.** Refusing Mode A, dropping a term, declining to score an unsourced\nweight \u2014 those are findings the run carries, not failures of the job. `failed` is\nreserved for the machinery breaking: Overpass returning 429, a missing token, a\nworker dying. Collapsing the two would put the engine's honesty into an error log\nwhere nobody reads it.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "rejected"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobSubmission": {
        "additionalProperties": false,
        "description": "One request to assess something. Exactly one of `corridor_id` or `panel`.\n\nTwo ways in, because the command line has two: `roadrisk corridor` builds a panel\nfrom geography, `roadrisk assess` judges one you already have. Neither is the\ndegraded case.",
        "properties": {
          "corridor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corridor Id"
          },
          "panel": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Rows of a panel you already built, as objects. Validated against the input contract before the job exists, so a panel that could never be assessed never becomes a queued job.",
            "title": "Panel"
          },
          "params": {
            "$ref": "#/components/schemas/JobOptions"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "JobSubmission",
        "type": "object"
      },
...
      "Refusal": {
        "additionalProperties": false,
        "description": "Why the request was refused, in the terms the caller can act on.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "errors": {
            "description": "Every fault found, when a request had more than one.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The single thing at fault, when there is one.",
            "title": "Field"
          },
          "message": {
            "description": "Written for a person. Names the column, the id or the setting.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Refusal",
        "type": "object"
      },
      "Region": {
        "description": "Where a weight was estimated, or where a corridor is.\n\nRegion granularity, not country, because published weights are estimated on\nregional or national datasets and never on \"Cyprus\" specifically. A Cyprus\ncorridor declares ``europe`` and gets European evidence where it exists, global\nevidence otherwise, and North American evidence only as a last resort \u2014 with the\nreach reported.\n\nStage 2 will resolve this from the corridor's admin boundary automatically; the\nGADM and OSM-relation adapters are already declared for that.",
        "enum": [
          "north_america",
          "europe",
          "australasia",
          "asia",
          "africa",
          "middle_east",
          "latin_america",
          "global"
        ],
        "title": "Region",
        "type": "string"
      },
...
      "Severity": {
        "description": "Which crashes a weight predicts.\n\nThis is not decoration. The Elvik Power Model exponent is 1.6 for injury crashes\nand 4.1 for fatal ones \u2014 applying the wrong one is a factor-of-two error, and\nbefore this existed the registry silently assumed injury.",
        "enum": [
          "all",
          "injury",
          "fsi",
          "fatal"
        ],
        "title": "Severity",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "\nCorridor road-risk assessment from open data, with provenance.\n\n**Two things this deployment does not yet do, stated here rather than discovered.**\n\n* **Nothing executes jobs.** `POST /jobs` stores a job and returns `202`. It stays\n  `queued` until step 5.1d attaches a runner. `GET /health` reports `runner: null`.\n* **`X-Tenant-Id` is not authentication.** It scopes every read to one tenant, which is\n  what keeps two clients' runs apart, but nothing verifies the claim. Step 5.4a\n  replaces it with real identities and row-level policies in the database.\n\n**A refusal is a result, not an error.** A panel that breaks the input contract is a\n`422` naming the column, and no job is created. An assessment that descended to Mode B,\ndropped a term or declined to score an unsourced weight is a **`200`** \u2014 those are\nfindings the run carries, not failures. Infrastructure breaking is a job whose status is\n`failed`, with a cause.\n\nPayload shape version `1.0`. Factors, tiers and licences come from\n`factors.yaml`; `GET /registry` serves it with the hash of the file it was read from.\n",
    "summary": "Corridor risk from open data, with every number traceable.",
    "title": "Road Risk Panel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "description": "Accept a job, or refuse it before it exists.\n\nNote the order. Every check that can refuse happens *before* `create_job`, so a\n422 leaves no row behind \u2014 \"no job created\" is the half of the refusal contract\nthat a client can actually verify, by listing the project's jobs and finding it\nempty.",
        "operationId": "submit_job_jobs_post",
        "parameters": [
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Queued. Poll `Location`, which is `GET /jobs/{id}`."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "413": {
            "description": "The panel is larger than this deployment accepts."
          },
          "422": {
            "description": "The panel broke the input contract, or the request could not describe a runnable job. No job was created."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Submit an assessment. Accepted, not performed",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "description": "Read a job's status.\n\n`succeeded` means the job ran to completion, **not** that Mode A was reached. A run\nthat descended to Mode B, dropped a term or refused an unsourced weight succeeded,\nand carries those findings. `failed` is the machinery breaking, with a cause in\n`error` and never a stack trace. `rejected` is the panel breaking the input\ncontract, where nothing malfunctioned.\n\nAt 5.1c nothing executes jobs, so every job here stays `queued`. `GET /health`\nreports `runner: null` for exactly that reason.",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Where a job got to",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/projects/{project_id}/jobs": {
      "get": {
        "operationId": "list_jobs_projects__project_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Jobs Projects  Project Id  Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Every job in a project, newest first",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}